befor ordering

+++ lib/page/home/forms.dart
@@ -21,6 +21,62 @@ class _ProductFormPageState extends State<ProductFormPage> {
   void initState() {
     super.initState();
     _fetchForms();
+    _fetchProducts();
+  }
+
+  // Update a product form with new name
+  Future<void> _updateForm(String formId, String name) async {
+    var prefs = await SharedPreferences.getInstance();
+    String? accessToken = prefs.getString('access_token');
+    final response = await http.put(
+      Uri.parse('https://smset.ir/product/api/v1/product_forms/$formId/'),
+      headers: {
+        'Authorization': 'Bearer $accessToken',
+        'Content-Type': 'application/json',
+      },
+      body: json.encode({'name': name}),
+    );
+
+    if (response.statusCode == 200) {
+      _fetchForms(); // Refresh the forms list
+    } else {
+      throw Exception('Failed to update form');
+    }
+  }
+
+  Future<void> _editForm(String formId, String currentName) async {
+    final TextEditingController nameController =
+        TextEditingController(text: currentName);
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text('ویرایش فرم محصول'),
+          content: TextField(
+            controller: nameController,
+            decoration: const InputDecoration(labelText: 'نام فرم'),
+          ),
+          actions: [
+            TextButton(
+              onPressed: () {
+                Navigator.pop(context);
+              },
+              child: const Text('انصراف'),
+            ),
+            TextButton(
+              onPressed: () async {
+                final formName = nameController.text;
+                if (formName.isEmpty) return;
+
+                await _updateForm(formId, formName);
+                Navigator.pop(context);
+              },
+              child: const Text('ویرایش'),
+            ),
+          ],
+        );
+      },
+    );
   }
 
   // Fetch forms from the API
@@ -49,18 +105,18 @@ class _ProductFormPageState extends State<ProductFormPage> {
       setState(() {
         _isLoading = false;
       });
-      throw Exception('Failed to load forms');
+      throw Exception('${json.decode(response.body)}');
     }
   }
 
-  // Fetch products for the current user
+  // Fetch products from the API
   Future<void> _fetchProducts() async {
     setState(() {
       _isLoadingProducts = true;
     });
+
     var prefs = await SharedPreferences.getInstance();
     String? accessToken = prefs.getString('access_token');
-
     final response = await http.get(
       Uri.parse('https://smset.ir/product/api/v1/product/'),
       headers: {
@@ -79,13 +135,27 @@ class _ProductFormPageState extends State<ProductFormPage> {
       setState(() {
         _isLoadingProducts = false;
       });
-      throw Exception('Failed to load products');
+      throw Exception('Failed to fetch products');
     }
   }
 
-  Future<void> _showProductSelectionDialog(String formId) async {
-    // ابتدا محصولات را بارگذاری می‌کنیم
-    await _fetchProducts();
+  Future<void> _showProductSelectionDialog(formId) async {
+    print("products $_products");
+    for (var i = 0; i < formId['items'].length; i++) {
+      bool isSelected = _selectedProducts
+          .any((item) => item['id'] == formId['items'][i]['id']);
+      if (!isSelected){
+        var productId = formId['items'][i];
+        print("productId $productId");
+        for (var x in _products) {
+          if (x['id']==productId['product']){
+            _selectedProducts.add(x);
+          }
+        }
+      }
+    }
+    print(formId);
+    // await _fetchProducts();
     showDialog(
       context: context,
       builder: (context) {
@@ -93,36 +163,36 @@ class _ProductFormPageState extends State<ProductFormPage> {
           title: const Text('Select Products'),
           content: _isLoadingProducts
               ? const Center(child: CircularProgressIndicator())
-              : Container(
-            height: 290,
-            width: 190,
-            child: StatefulBuilder(
-              builder: (BuildContext context, setState) {
-                return ListView.builder(
-                  itemCount: _products.length,
-                  itemBuilder: (context, index) {
-                    final product = _products[index];
-                    bool isSelected = _selectedProducts
-                        .any((item) => item['id'] == product['id']);
-                    return CheckboxListTile(
-                      title: Text(product['name']),
-                      value: isSelected,
-                      onChanged: (bool? selected) {
-                        setState(() {
-                          if (selected == true) {
-                            _selectedProducts.add(product);
-                          } else {
-                            _selectedProducts.removeWhere(
-                                    (item) => item['id'] == product['id']);
-                          }
-                        });
-                      },
-                    );
-                  },
-                );
-              },
-            ),
-          ),
+              : SizedBox(
+                  height: 290,
+                  width: 190,
+                  child: StatefulBuilder(
+                    builder: (BuildContext context, setState) {
+                      return ListView.builder(
+                        itemCount: _products.length,
+                        itemBuilder: (context, index) {
+                          final product = _products[index];
+                          bool isSelected = _selectedProducts
+                              .any((item) => item['id'] == product['id']);
+                          return CheckboxListTile(
+                            title: Text(product['name']),
+                            value: isSelected,
+                            onChanged: (bool? selected) {
+                              setState(() {
+                                if (selected == true) {
+                                  _selectedProducts.add(product);
+                                } else {
+                                  _selectedProducts.removeWhere(
+                                      (item) => item['id'] == product['id']);
+                                }
+                              });
+                            },
+                          );
+                        },
+                      );
+                    },
+                  ),
+                ),
           actions: [
             TextButton(
               onPressed: () {
@@ -132,8 +202,8 @@ class _ProductFormPageState extends State<ProductFormPage> {
             ),
             TextButton(
               onPressed: () async {
-                // ارسال محصولات انتخابی به فرم
-                await _addProductsToForm(formId);
+                // Send selected products to the form
+                await _addProductsToForm(formId['id']);
                 Navigator.pop(context);
               },
               child: const Text('Save'),
@@ -144,13 +214,14 @@ class _ProductFormPageState extends State<ProductFormPage> {
     );
   }
 
+  // Add selected products to form
   Future<void> _addProductsToForm(String formId) async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
     String? accessToken = prefs.getString('access_token');
 
-    // استخراج شناسه‌های محصولات از لیست انتخاب شده
+    // Extract product IDs from selected products
     final List<String> productIds =
-    _selectedProducts.map((product) => product['id'].toString()).toList();
+        _selectedProducts.map((product) => product['id'].toString()).toList();
 
     final response = await http.post(
       Uri.parse(
@@ -163,48 +234,71 @@ class _ProductFormPageState extends State<ProductFormPage> {
     );
 
     if (response.statusCode == 201) {
-      // پس از ارسال موفق، لیست فرم‌ها را دوباره بکشیم
-      _fetchForms(); // Refresh the forms list
+      _fetchForms(); // Refresh the forms list after adding products
     } else {
-      throw Exception('Failed to add products');
+      throw Exception('Failed to add products ${ response.body}response.statusCode');
     }
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text("Product Forms")),
+      appBar: AppBar(
+        title: const Text(
+          "فرم‌های محصولات",
+          style: TextStyle(fontFamily: 'Vazir', fontWeight: FontWeight.bold),
+        ),
+        backgroundColor: Colors.blue,
+      ),
       body: _isLoading
           ? const Center(child: CircularProgressIndicator())
           : ListView.builder(
               itemCount: _forms.length,
               itemBuilder: (context, index) {
                 final form = _forms[index];
-                return ListTile(
-                  title: Text(form['name']),
-                  subtitle: Text('ID: ${form['id']}'),
-                  trailing: Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      IconButton(
-                        icon: const Icon(Icons.edit),
-                        onPressed: () {
-                          // Implement form edit functionality if needed
-                        },
-                      ),
-                      IconButton(
-                        icon: const Icon(Icons.delete),
-                        onPressed: () {
-                          _deleteForm(form['id']);
-                        },
-                      ),
-                      IconButton(
-                        icon: const Icon(Icons.add),
-                        onPressed: () {
-                          _showProductSelectionDialog(form['id']);
-                        },
-                      ),
-                    ],
+                return Card(
+                  margin:
+                      const EdgeInsets.symmetric(vertical: 8, horizontal: 10),
+                  elevation: 4,
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(12),
+                  ),
+                  child: ListTile(
+                    contentPadding: const EdgeInsets.symmetric(
+                        vertical: 12, horizontal: 16),
+                    title: Text(
+                      form['name'],
+                      style: const TextStyle(
+                          fontWeight: FontWeight.bold, fontSize: 16),
+                    ),
+                    subtitle:
+                        const Text('برای ویرایش یا مدیریت محصولات ضربه بزنید'),
+                    trailing: Row(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        IconButton(
+                          icon: const Icon(Icons.edit, color: Colors.blue),
+                          onPressed: () {
+                            _editForm(form['id'], form['name']);
+                            // Add your edit functionality here
+                          },
+                        ),
+                        IconButton(
+                          icon: const Icon(Icons.delete, color: Colors.red),
+                          onPressed: () {
+                            _deleteForm(form[
+                                'id']); // Call delete method with confirmation
+                          },
+                        ),
+                        IconButton(
+                          icon: const Icon(Icons.add, color: Colors.green),
+                          onPressed: () {
+                            _showProductSelectionDialog(
+                                form); // Show product selection dialog
+                          },
+                        ),
+                      ],
+                    ),
                   ),
                 );
               },
@@ -216,17 +310,17 @@ class _ProductFormPageState extends State<ProductFormPage> {
             context: context,
             builder: (context) {
               return AlertDialog(
-                title: const Text('Create Product Form'),
+                title: const Text('ایجاد فرم محصول'),
                 content: TextField(
                   controller: nameController,
-                  decoration: const InputDecoration(labelText: 'Form Name'),
+                  decoration: const InputDecoration(labelText: 'نام فرم'),
                 ),
                 actions: [
                   TextButton(
                     onPressed: () {
                       Navigator.pop(context);
                     },
-                    child: const Text('Cancel'),
+                    child: const Text('انصراف'),
                   ),
                   TextButton(
                     onPressed: () async {
@@ -236,7 +330,7 @@ class _ProductFormPageState extends State<ProductFormPage> {
                       await _createForm(formName);
                       Navigator.pop(context);
                     },
-                    child: const Text('Create'),
+                    child: const Text('ایجاد'),
                   ),
                 ],
               );
@@ -244,6 +338,7 @@ class _ProductFormPageState extends State<ProductFormPage> {
           );
         },
         child: const Icon(Icons.add),
+        backgroundColor: Colors.blue,
       ),
     );
   }
@@ -268,21 +363,46 @@ class _ProductFormPageState extends State<ProductFormPage> {
     }
   }
 
-  // Delete a product form
+  // Delete a product form with confirmation
   Future<void> _deleteForm(String formId) async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    String? accessToken = prefs.getString('access_token');
-    final response = await http.delete(
-      Uri.parse('https://smset.ir/product/api/v1/product_forms/$formId/'),
-      headers: {
-        'Authorization': 'Bearer $accessToken',
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text('آیا مطمئن هستید؟'),
+          content: const Text('آیا از حذف این فرم اطمینان دارید؟'),
+          actions: [
+            TextButton(
+              onPressed: () {
+                Navigator.pop(context);
+              },
+              child: const Text('انصراف'),
+            ),
+            TextButton(
+              onPressed: () async {
+                SharedPreferences prefs = await SharedPreferences.getInstance();
+                String? accessToken = prefs.getString('access_token');
+                final response = await http.delete(
+                  Uri.parse(
+                      'https://smset.ir/product/api/v1/product_forms/$formId/'),
+                  headers: {
+                    'Authorization': 'Bearer $accessToken',
+                  },
+                );
+
+                if (response.statusCode == 204) {
+                  Navigator.pop(context);
+                  _fetchForms(); // Refresh the forms list after deletion
+                } else {
+                  Navigator.pop(context);
+                  throw Exception('Failed to delete form');
+                }
+              },
+              child: const Text('حذف'),
+            ),
+          ],
+        );
       },
     );
-
-    if (response.statusCode == 204) {
-      _fetchForms(); // Refresh the forms list
-    } else {
-      throw Exception('Failed to delete form');
-    }
   }
 }
